Update package.json with relevant information
Extended the package.json with detailed information like description, version update, author details, repository details, and relevant keywords. The `1.0.1` version adds all this information for users and other developers. It provides a clearer understanding of the package, its usage, and the author. It also helps with SEO on npm registry. The output file of the build process is also changed from `index.mjs` to `index.js` for better compatibility.